+++ src/FreeElephants/JsonApiToolkit/DTO/RelationshipToOne.php
@@ -2,23 +2,7 @@
 
 namespace FreeElephants\JsonApiToolkit\DTO;
 
-use Psr\Http\Message\ServerRequestInterface;
-
-class RelationshipToOne
+class RelationshipToOne extends AbstractDocument
 {
     public ResourceIdentifierObject $data;
-
-    public function __construct($data)
-    {
-        $this->data = new ResourceIdentifierObject($data['data']);
-    }
-
-    public static function fromRequest(ServerRequestInterface $request): self
-    {
-        $request->getBody()->rewind();
-        $rawJson = $request->getBody()->getContents();
-        $decodedJson = json_decode($rawJson, true);
-
-        return new self($decodedJson);
-    }
 }